Enemy health bars, saving loading, weapon/armor rework.
Implemented enemy health bars, can probably be optimised for only when
they should be on screen, currently run all the time.

Saving and loading sort of fully works, the only problem now (as far as
I can see) is that the UI updating is funky (gold update is weird on
load).